requirements-oss.txt
PyYAML==5.3.1
rasa==2.1.3
spacy==2.3.4
rasa[spacy]